refactor: 역할에 따라 Track 관련 코드 폴더 분리
