+++ src/main/java/com/petitapetit/miml/domain/notification/event/TrackAddedEvent.java
@@ -1,6 +1,6 @@
 package com.petitapetit.miml.domain.notification.event;
 
-import com.petitapetit.miml.domain.track.Track;
+import com.petitapetit.miml.domain.track.entity.Track;
 
 // 신곡이 추가될 때 발행(publish) 될 event
 public class TrackAddedEvent {
