+++ src/main/java/com/petitapetit/miml/domain/track/entity/Track.java
@@ -1,6 +1,7 @@
-package com.petitapetit.miml.domain.track;
+package com.petitapetit.miml.domain.track.entity;
 
 import com.petitapetit.miml.domain.artist.domain.Artist;
+import com.petitapetit.miml.domain.track.dto.TrackDto;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
