+++ src/main/java/com/petitapetit/miml/domain/track/repository/ArtistTrackRepository.java
@@ -1,5 +1,6 @@
-package com.petitapetit.miml.domain.track;
+package com.petitapetit.miml.domain.track.repository;
 
+import com.petitapetit.miml.domain.track.entity.ArtistTrack;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistTrackRepository extends JpaRepository<ArtistTrack, Long> {
