+++ src/test/java/com/petitapetit/miml/domain/notification/NotificationEventHandlerTest.java
@@ -15,8 +15,8 @@ import com.petitapetit.miml.domain.notification.event.FriendRequestedEvent;
 import com.petitapetit.miml.domain.notification.service.NotificationEventHandler;
 import com.petitapetit.miml.domain.notification.event.SharePlaylistRequestedEvent;
 import com.petitapetit.miml.domain.notification.event.TrackAddedEvent;
-import com.petitapetit.miml.domain.track.Track;
-import com.petitapetit.miml.domain.track.TrackDto;
+import com.petitapetit.miml.domain.track.entity.Track;
+import com.petitapetit.miml.domain.track.dto.TrackDto;
 import com.petitapetit.miml.test.ServiceTest;
 import java.util.ArrayList;
 import java.util.List;
