+++ src/test/java/com/petitapetit/miml/domain/notification/NotificationTransactionIntegrationTest.java
@@ -11,12 +11,12 @@ import com.petitapetit.miml.domain.member.model.RoleType;
 import com.petitapetit.miml.domain.member.repository.MemberRepository;
 import com.petitapetit.miml.domain.notification.entity.Notification;
 import com.petitapetit.miml.domain.notification.repository.NotificationRepository;
-import com.petitapetit.miml.domain.track.ArtistTrack;
-import com.petitapetit.miml.domain.track.ArtistTrackRepository;
-import com.petitapetit.miml.domain.track.Track;
-import com.petitapetit.miml.domain.track.TrackDto;
-import com.petitapetit.miml.domain.track.TrackRepository;
-import com.petitapetit.miml.domain.track.TrackService;
+import com.petitapetit.miml.domain.track.entity.ArtistTrack;
+import com.petitapetit.miml.domain.track.repository.ArtistTrackRepository;
+import com.petitapetit.miml.domain.track.entity.Track;
+import com.petitapetit.miml.domain.track.dto.TrackDto;
+import com.petitapetit.miml.domain.track.repository.TrackRepository;
+import com.petitapetit.miml.domain.track.service.TrackService;
 import java.util.ArrayList;
 import java.util.Set;
 import java.util.stream.Collectors;
